feat: Add follow logic test code

+++ src/resolvers/app/follow/Mutation.ts
@@ -1,24 +1,28 @@
 import { Context } from "config/types"
 import { FollowInput } from "resolvers/app/follow/models"
+import { ApolloError } from "apollo-server-express"
 
 /**
  * 팔로우 상태 반전
  * @param args.input.following        내가 팔로우할 사람  
  */
 export const follow = async (parent: void, args: FollowInput, context: Context) => {
+    if (context.user.id === args.input.following) {
+        return new ApolloError("자기 자신을 팔로잉할 수 없습니다")
+    }
     const result = await context.db.collection("follow").find({
-        userId: args.input.following,
-        following: context.user.id
+        userId: context.user.id,
+        following: args.input.following
     }).count()
     if (result === 0) {
         return context.db.collection("follow").insertOne({
-            userId: args.input.following,
-            following: context.user.id
+            userId: context.user.id,
+            following: args.input.following
         }).then(({ result }) => result.n === 1)
     } else {
         return context.db.collection("follow").deleteOne({
-            userId: args.input.following,
-            following: context.user.id
+            userId: context.user.id,
+            following: args.input.following
         }).then(({ result }) => result.n === 1)
     }
 }
